+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "editor.indentSize": 4,
+  "editor.formatOnSave": false
+}
