+++ constants.js
@@ -1,0 +1,16 @@
+//////////// CONSTANTS ///////////////
+export const FRAMERATE = 60; // DO NOT MODIFY
+export const SHIP_SPEED = 1;
+export const SIM_FACTOR = 2;
+export const NUM_PLANETS = 5;
+export const WORLD_SEED = 7;
+export const COMBAT_RATE = 20;
+export const BASE_PRODUCTION_RATE = 1;
+export const BASE_TERRAFORMING_RATE = 500; // smaller is faster
+export const LEVEL_NUM = 3;
+
+export const teamColorMap = {
+  BLUE: `hsl(${0.55 * 360}, 80%, 60%)`,
+  RED: `hsl(${0.05 * 360}, 80%, 60%)`,
+  YELLOW: `hsl(${0.35 * 360}, 80%, 60%)`,
+};
