+++ index.html
@@ -53,7 +53,7 @@
     <canvas id="gameCanvas"></canvas>
     <div id="buttonContainer">
   </div>
-  <script src="main.js"></script>
+  <script type="module" src="main.js"></script>
 </body>
 </html>
